properly feeding state
